Abstract out example goal component
