+++ src/views/Home.js
@@ -10,6 +10,7 @@ import { changeView } from "../state/actions/router"
 import { selectTopic, setPageIndex } from "../state/actions/tips"
 
 import NavBar from "../components/NavBar"
+import ExampleGoal from "../components/ExampleGoal"
 
 import { Tips, NewFlow, InternalLink, ExternalLink } from "../Constants"
 
@@ -129,34 +130,6 @@ const _OptionContainer = styled.div.attrs({
   flex-direction: ${({ number }) => (number === 2 ? "row" : "column")};
   flex-shrink: 0;
 `
-
-const _monoText = styled.p.attrs({ className: "mono db" })``
-const _sansText = styled.p.attrs({ className: "sans fw5 font-4 db pb2" })`
-  font-size: 16px;
-`
-const _exampleGoalDiv = styled.div.attrs({
-  className: "tc font-5 db pt3 pb2 ph3 mb2 bg-white dark-gray",
-})`
-  border-radius: 21px 21px 21px 6px;
-  max-width: 265px;
-  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.1);
-`
-const ExampleGoal = () => (
-  <_exampleGoalDiv>
-    <_monoText>Your goal is</_monoText>
-    <_sansText>
-      Stop procrasting and work on my time management skills
-    </_sansText>
-    <_monoText>by using</_monoText>
-    <_sansText className="underline">Procrastination tips</_sansText>
-    <_monoText>and you will do it</_monoText>
-    <_sansText>Every Tuesday and Friday</_sansText>
-    <_monoText>preferably</_monoText>
-    <_sansText>After school, at 16:30</_sansText>
-    <_monoText>for</_monoText>
-    <_sansText>3 months</_sansText>
-  </_exampleGoalDiv>
-)
 
 const RenderConversation = ({ conversation, onLinkClick, userAvatar }) =>
   conversation.map(({ content, type }, i) => {
